+++ src/main/kotlin/trackerr/rezyfr/dev/model/Transaction.kt
@@ -1,9 +1,8 @@
 package trackerr.rezyfr.dev.model
 
-import java.math.BigDecimal
 
 data class Transaction(
-    val amount: BigDecimal,
+    val amount: Double,
     val description: String,
     val categoryId: Int,
     val walletId: Int,
